refactor(cli): split doctor diagnostics
